Server prototype non modular (#5)
* Prototype of a vertx HTTP server for strimzi-admin
that is extensible through the Java SPI.

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

* Initial server infrastructure

Contains an initial implementation of the Strimzi Admin server.

The main application is in the `http-server` module. It creates
a vertx verticle which loads routes from modules on the classpath
using the Java service loader and then opens port 8080 to listen
for requests.

There are currently 2 modules containing routes. They are the
`health` module and the `graphql` module. The `health` module
contains a `liveness` and a `status` resource which return a
static `status:ok` json response. The `graphql` module implements
the graphql server. The graphql server creates a base server
with no queries or mutations. Currently, only sample queries are
implemented. The queries are defined in business log modules of
which the `kafka-admin` module is defined as an example. The
graphql server uses the Java service loader as well to register
graphql schema and implementation details to the graphql server.
If the `VERTXWEB_ENVIRONMENT` envar is set to `dev`, the graphql
server will also serve `GraphiQL` as per the `vertx-web-graphql`
documentation.

To server can be built using `mvn clean package` and can then
be run using the following:

java -cp kafka-admin/target/kafka-admin-1.0-SNAPSHOT-fat.jar:health/target/health-1.0-SNAPSHOT-fat.jar:grapql/target/graphql-s.0-SNAPSHOT-far.jar:http-server/target/http-server-1.0-SNAPSHOT-fat.jar io.strimzi.admin.Main

Tests are to follow.

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

* Updates as per review comments

* Added the maven-dependency-plugin `Analyze` rules and tidied the pom files.
* Added try/catch blocks for i/o operations in the executeBlocking blocks.
* Added Javadoc to main classes.
* Removed the duplicated sample data and made the data publicly visible so that
both the TopicHandler and TopicListHandler could access it.
* Uppercased the name of the sample data as it is a constant

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>